bike_sell
Consider the following scenario.

"Rahul is a PhD student at IIITDMJ who has recently completed his PhD program and he is leaving married hostel. He takes a decision that he should sell his old bike. He shares this message to his colleagues. Mohan is interested in buying the Rahul's bike but he is confused with the value and the condition of bike. He hires Abdul who is a bike mechanic for checking the condition of bike. Mohan decided that if Abdul says that bike is in good condition, then he will ask Neeraj for valuation of bike to get the actual cost as per the condition.  He puts the final price in-front of Rahul who is the bike owner. If both of them agree they will sign the contract."